Hex Grid
Early version of procedural hex grid is complete with random selection of base and topper.

Need to tighten the tiles as some float above the platform a bit which may be noticeable depending on camera.

Also need to add biases eventually to generate paths and obstacles logically.

Imported Synty Mini Fantasy assets to help test and it's my asset pack of choice for this project.